# internal/core_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A book management service API in Go using Gin framework..",
        "title": "Phygital Core Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://fygito.com//support",
            "email": "support@kardialab.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/scan/nfc/tap/{tag_id}": {
            "get": {
                "description": "nfc tap",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "NFC Tap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "tag_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "$ref": "#/definitions/handler.RedirectResponse"
                        }
                    },
                    "303": {
                        "description": "See Other",
                        "schema": {
                            "$ref": "#/definitions/handler.RedirectResponse"
                        }
                    }
                }
            }
        },
        "/scan/nfc/verify": {
            "get": {
                "description": "Nfc verify scan",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "NFC Verify Scan",
                "parameters": [
                    {
                        "type": "string",
                        "name": "cmac",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "enc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "picc_data",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "$ref": "#/definitions/handler.RedirectResponse"
                        }
                    },
                    "303": {
                        "description": "See Other",
                        "schema": {
                            "$ref": "#/definitions/handler.RedirectResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.RedirectResponse": {
            "type": "object",
            "properties": {
                "status_code": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}